pre-merge massage
